% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skygrowth.R
\name{continue.phylo.bnpma}
\alias{continue.phylo.bnpma}
\title{Continue MCMC chain of previously fitted model}
\usage{
continue.phylo.bnpma(fit, mhsteps, quiet = F, ...)
}
\arguments{
\item{fit}{A phylo.bnpma fit}

\item{mhsteps}{Additional MCMC steps to compute}

\item{quiet}{Verbose output?}
}
\value{
A phylo.bnpma fit
}
\description{
Continue MCMC chain of previously fitted model
}
