% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skygrowth.R
\name{continue.skygrowth.mcmc}
\alias{continue.skygrowth.mcmc}
\title{Continue MCMC chain of previously fitted model}
\usage{
continue.skygrowth.mcmc(fit, mhsteps, quiet = F, ...)
}
\arguments{
\item{fit}{A skygrowth.mcmc fit}

\item{mhsteps}{Additional MCMC steps to compute}

\item{quiet}{Verbose output?}

\item{...}{Passed on}
}
\value{
A skygrowth.mcmc fit
}
\description{
Continue MCMC chain of previously fitted model
}
