% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armasky.R
\name{growth.plot}
\alias{growth.plot}
\title{Plot growth rate of effective size through time}
\usage{
growth.plot(x, ...)
}
\arguments{
\item{fit}{A fitted object (eg phylo.mapma)}
}
\value{
A ggplot2 plot
}
\description{
Plot growth rate of effective size through time
}
