% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armasky.R
\name{ne.plot}
\alias{ne.plot}
\title{Plot effective size through time}
\usage{
ne.plot(x, ...)
}
\arguments{
\item{fit}{A fitted object (eg phylo.mapma)}
}
\value{
A ggplot2 plot
}
\description{
Plot effective size through time
}
