% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skygrowth.R
\name{neplot}
\alias{neplot}
\title{Plot effective size through time}
\usage{
neplot(x, ...)
}
\arguments{
\item{...}{Additional parameters are passed to ggplot or the base plotting function}

\item{fit}{A fitted object (eg skygrowth.map or skygrowth.mcmc)}

\item{logy=TRUE}{If TRUE, the plot is returned with logarithmic y-axis}

\item{ggplot=TRUE}{If TRUE, returns a ggplot2 figure}
}
\value{
A ggplot2 plot
}
\description{
Plot effective size through time
}
