% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armasky.R
\name{phylo.bnpma.covar}
\alias{phylo.bnpma.covar}
\title{A gibbs-metropolis algorithm for sampling Ne(t) with a 1st order moving average model and using covariate data}
\usage{
phylo.bnpma.covar(tre, formula, data, maxSampleTime, iter = 1e+05,
  iter0 = 10, tau0 = 10, tau_logprior = "exponential", res = 50,
  beta_logpriors = list(), prop_beta_sd = NULL, quiet = F,
  control = NULL, ...)
}
\arguments{
\item{tre}{A dated phylogeny in ape::phylo format (see documentation for ape)}

\item{formula}{An R formula with empty left-hand-side; the right-hand-side specifies relationship of covariates with growth rate of Ne
data A data.frame, must include 'time' column
maxSampleTime The scalar time that the most recent sample was collected}

\item{tau0}{Initial guess of the precision parameter}

\item{tau_logprior}{Prior for precision parameter (character string (gamma or exponential) or function)}

\item{res}{Number of time points (integer)}

\item{beta_logpriors}{Optional list of functions providing log density for coefficients (must correspond to data)}

\item{quiet}{Provide verbose output?}

\item{control}{List of options passed to optim}

\item{mhsteps}{Number of mcmc steps}

\item{abstol}{Criterion for convergence of likelihood}
}
\value{
A fitted model including effective size through time
}
\description{
A gibbs-metropolis algorithm for sampling Ne(t) with a 1st order moving average model and using covariate data
}
