% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armasky.R
\name{R.plot}
\alias{R.plot}
\title{Plot reproduction number through time}
\usage{
R.plot(x, ...)
}
\arguments{
\item{...}{Additional parameters are passed to ggplot or the base plotting function}

\item{fit}{A fitted object (eg phylo.mapma)}

\item{ggplot=TRUE}{If TRUE, returns a ggplot2 figure}
}
\value{
A ggplot2 plot
}
\description{
Plot reproduction number through time
}
