% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skygrowth.R
\name{skygrowth.map.covar}
\alias{skygrowth.map.covar}
\title{Maximum a posteriori estimate of effective size through time using covariate data}
\usage{
skygrowth.map.covar(
  tre,
  formula,
  data,
  maxSampleTime,
  tau0 = 10,
  tau_logprior = "exponential",
  beta_logpriors = list(),
  res = 50,
  quiet = FALSE,
  maxiter = 20,
  abstol = 0.01,
  control = NULL,
  maxHeight = NULL,
  ...
)
}
\arguments{
\item{tre}{A dated phylogeny in ape::phylo format (see documentation for ape) or a list of phylogenies or multi.phylo object}

\item{formula}{An R formula with empty left-hand-side; the right-hand-side specifies relationship of covariates with growth rate of Ne}

\item{data}{A data.frame, must include 'time' column}

\item{maxSampleTime}{The scalar time that the most recent sample was collected}

\item{tau0}{Initial guess of the precision parameter}

\item{tau_logprior}{Prior for precision parameter (character string (gamma or exponential) or function)}

\item{beta_logpriors}{Optional list of functions providing log density for coefficients (must correspond to data)}

\item{res}{Number of time points (integer)}

\item{quiet}{Provide verbose output?}

\item{maxiter}{Maximum number of iterations}

\item{abstol}{Criterion for convergence of likelihood}

\item{control}{List of options passed to optim}

\item{maxHeight}{The earliest time (furthest time in to the past) to estimate Ne}

\item{...}{Not implemented}
}
\value{
A fitted model including effective size through time
}
\description{
Maximum a posteriori estimate of effective size through time using covariate data
}
