% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skygrowth.R
\name{skygrowth.map}
\alias{skygrowth.map}
\title{Maximum a posteriori estimate of effective size through time with a non-parametric growth model}
\usage{
skygrowth.map(tre, tau0 = 10, tau_logprior = "exponential", res = 50,
  quiet = FALSE, maxiter = 20, abstol = 0.01, control = NULL, ...)
}
\arguments{
\item{tre}{A dated phylogeny in ape::phylo format (see documentation for ape)}

\item{tau0}{Initial guess of the precision parameter}

\item{tau_logprior}{Prior for precision parameter (character string (gamma or exponential) or function)}

\item{res}{Number of time points (integer)}

\item{quiet}{Provide verbose output?}

\item{maxiter}{Maximum number of iterations}

\item{abstol}{Criterion for convergence of likelihood}

\item{control}{List of options passed to optim}

\item{...}{Not implemented}
}
\value{
A fitted model including effective size through time
}
\description{
Maximum a posteriori estimate of effective size through time with a non-parametric growth model
}
