% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skygrowth.R
\name{skygrowth.mcmc.covar}
\alias{skygrowth.mcmc.covar}
\title{A gibbs-metropolis algorithm for sampling Ne(t) with a 1st order moving average model and using covariate data}
\usage{
skygrowth.mcmc.covar(
  tre,
  formula,
  data,
  maxSampleTime,
  iter = 1e+05,
  iter0 = 10,
  tau0 = 10,
  tau_logprior = "exponential",
  res = 50,
  beta_logpriors = list(),
  prop_beta_sd = NULL,
  quiet = FALSE,
  control = NULL,
  gamma = NA,
  logRmean = NULL,
  logRsd = 1,
  maxHeight = NULL
)
}
\arguments{
\item{tre}{A dated phylogeny in ape::phylo format (see documentation for ape)}

\item{formula}{An R formula with empty left-hand-side; the right-hand-side specifies relationship of covariates with growth rate of Ne}

\item{data}{A data.frame, must include 'time' column}

\item{maxSampleTime}{The scalar time that the most recent sample was collected}

\item{iter}{iter}

\item{iter0}{iter0}

\item{tau0}{Initial guess of the precision parameter}

\item{tau_logprior}{Prior for precision parameter (character string (gamma or exponential) or function)}

\item{res}{Number of time points (integer)}

\item{beta_logpriors}{Optional list of functions providing log density for coefficients (must correspond to data)}

\item{prop_beta_sd}{Standard deviation of beta proposal kernel}

\item{quiet}{Provide verbose output?}

\item{control}{List of options passed to optim}

\item{gamma}{Death rate. If provided will compute R}

\item{logRmean}{Mean of R in log space. Determines a lognormal prior on R(t). If used, _gamma_ must be provided}

\item{logRsd}{SD of R in log space. Determines a lognormal prior on R(t). If used, _gamma_ must be provided}

\item{maxHeight}{If supplied, will only compute Ne(t) estimates this far back in time. Otherwise will compute to the root of the tree.}

\item{mhsteps}{Number of mcmc steps}
}
\value{
A fitted model including effective size through time
}
\description{
A gibbs-metropolis algorithm for sampling Ne(t) with a 1st order moving average model and using covariate data
}
