% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skygrowth.R
\name{growth.plot}
\alias{growth.plot}
\title{Plot growth rate of effective size through time}
\usage{
growth.plot(fit, ggplot = TRUE, logy = FALSE, ...)
}
\arguments{
\item{fit}{A fitted object (eg skygrowth.map)}

\item{ggplot}{If TRUE, returns a ggplot2 figure}

\item{logy}{If TRUE, the plot is returned with logarithmic y-axis}

\item{...}{Additional parameters are passed to ggplot or the base plotting function}
}
\value{
A ggplot2 plot
}
\description{
Plot growth rate of effective size through time
}
