% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armasky.R
\name{growth.plot}
\alias{growth.plot}
\title{Plot growth rate of effective size through time}
\usage{
growth.plot(x, ...)
}
\arguments{
\item{...}{Additional parameters are passed to ggplot or the base plotting function}

\item{fit}{A fitted object (eg phylo.mapma)}

\item{logy=TRUE}{If TRUE, the plot is returned with logarithmic y-axis}

\item{ggplot=TRUE}{If TRUE, returns a ggplot2 figure}
}
\value{
A ggplot2 plot
}
\description{
Plot growth rate of effective size through time
}
