% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skygrowth.R
\name{neplot}
\alias{neplot}
\title{Plot effective size through time}
\usage{
neplot(fit, ggplot = TRUE, logy = TRUE, ...)
}
\arguments{
\item{fit}{A fitted object (eg skygrowth.map or skygrowth.mcmc)}

\item{ggplot}{If TRUE, returns a ggplot2 figure}

\item{logy}{If TRUE, the plot is returned with logarithmic y-axis}

\item{...}{Additional parameters are passed to ggplot or the base plotting function}
}
\value{
A ggplot2 plot
}
\description{
Plot effective size through time
}
