% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skygrowth.R
\name{R.plot}
\alias{R.plot}
\title{Plot reproduction number through time}
\usage{
R.plot(fit, gamma = NA, ggplot = TRUE, ...)
}
\arguments{
\item{fit}{A fitted object (eg skygrowth.map)}

\item{gamma}{Value of gamma}

\item{ggplot}{If TRUE, returns a ggplot2 figure}

\item{...}{Additional parameters are passed to ggplot or the base plotting function}
}
\value{
A ggplot2 plot
}
\description{
Plot reproduction number through time
}
