% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skygrowth.R
\name{skygrowth.mcmc}
\alias{skygrowth.mcmc}
\title{A gibbs-metropolis algorithm for sampling Ne(t) with a non-parametric growth model}
\usage{
skygrowth.mcmc(tre, mhsteps = 1e+05, tau0 = 10,
  tau_logprior = "exponential", res = 50, quiet = F,
  control = NULL, ...)
}
\arguments{
\item{tre}{A dated phylogeny in ape::phylo format (see documentation for ape)}

\item{mhsteps}{Number of mcmc steps}

\item{tau0}{Initial guess of the precision parameter}

\item{tau_logprior}{Prior for precision parameter (character string (gamma or exponential) or function)}

\item{res}{Number of time points (integer)}

\item{quiet}{Provide verbose output?}

\item{control}{List of options passed to optim}

\item{...}{Not implemented}
}
\value{
A fitted model including effective size through time
}
\description{
A gibbs-metropolis algorithm for sampling Ne(t) with a non-parametric growth model
}
